requirements.txt
ccxt
ccxtpro
python-dotenv
websockets
pybluez
